Mise a jour des librairies
Mise à jour des chemins
corrections, ajout de champs lors les lectures de la base
