+++ lib/GaletteObjectsLend/Repository/LendObject.php
@@ -38,7 +38,7 @@
  * @since     Available since 0.7
  */
 
-namespace GaletteObjectsLend;
+namespace GaletteObjectsLend\Repository;
 
 use Analog\Analog;
 use \Zend\Db\Sql\Predicate;
@@ -78,12 +78,14 @@ class LendObject
     private $weight = 0.0;
     private $is_active = true;
     private $category_id;
+    private $category_name;
     private $nb_available = 1;
     // Requête sur le dernier statut de l'objet
     private $date_begin;
     private $date_forecast;
     private $date_end;
     private $status_text;
+    private $status_id;
     private $is_home_location = true;
     // Requête sur l'adhérent associé au statut
     private $nom_adh = '';
@@ -91,8 +93,7 @@ class LendObject
     private $email_adh = '';
     private $id_adh;
     private $rent_id;
-    private $in_stock;
-
+	private $comments;
     private $currency = '€';
     private $picture;
     private $cat_active = true;
@@ -189,28 +190,39 @@ class LendObject
         $this->nb_available = $r->nb_available;
         $this->rent_id = $r->rent_id;
 
-        //load last rent infos (status, member, and so on
-        if ($this->rent_id) {
-            if (property_exists($r, 'status_text')) {
-                $this->status_text = $r->status_text;
-            }
+		if (property_exists($r, 'status_text')) {
+			$this->status_text = $r->status_text;
+		}
+		if (property_exists($r, 'status_id')) {
+			$this->status_id = $r->status_id;
+		}
+		if (property_exists($r, 'comments')) {
+			$this->comments = $r->comments;
+		}
+		if (property_exists($r, 'date_begin')) {
+			$this->date_begin = $r->date_begin;
+		}
 
-            if (property_exists($r, 'date_begin')) {
-                $this->date_begin = $r->date_begin;
-            }
+		if (property_exists($r, 'date_forecast')) {
+			$this->date_forecast = $r->date_forecast;
+		}
 
-            if (property_exists($r, 'date_forecast')) {
-                $this->date_forecast = $r->date_forecast;
-            }
+		if (property_exists($r, 'date_end')) {
+			$this->date_end = $r->date_end;
+		}
 
-            if (property_exists($r, Adherent::PK)) {
-                $this->id_adh = $r->{Adherent::PK};
-            }
+		if (property_exists($r, Adherent::PK)) {
+			$this->id_adh = $r->{Adherent::PK};
+		}
 
-            if (property_exists($r, 'is_home_location')) {
-                $this->in_stock = $r->is_home_location;
-            }
-        }
+		if ($r->is_home_location == null) {
+			$this->is_home_location = true;
+		} else {
+				$this->is_home_location = $r->is_home_location;
+		}
+		$this->category_id = $r->category_id;
+		$this->category_name = $r->name;
+
 
         if ($this->object_id && $this->deps['rents'] === true) {
             $only_last = false;
@@ -315,6 +327,7 @@ class LendObject
             $object->date_forecast = $rent->date_forecast;
             $object->date_end = $rent->date_end;
             $object->status_text = $rent->status_text;
+			$object->comments = $rent->comments;
             $object->is_home_location = $rent->is_home_location == '1' ? true : false;
             $object->nom_adh = $rent->nom_adh;
             $object->prenom_adh = $rent->prenom_adh;
@@ -391,12 +404,7 @@ class LendObject
                 }
                 $dtb = new \DateTime($this->date_begin);
                 return $dtb->format('d/m/Y');
-            case 'date_end_ihm':
-                if ($this->date_end == '' || $this->date_end == null) {
-                    return '';
-                }
-                $dtb = new \DateTime($this->date_end);
-                return $dtb->format('j M Y');
+
             case 'date_forecast_ihm':
                 if ($this->date_forecast == '' || $this->date_forecast == null) {
                     return '';
@@ -499,25 +507,21 @@ class LendObject
             case 'name':
                 if ($filters->field_filter == Objects::FILTER_NAME) {
                     $process = true;
-                    continue;
                 }
                 break;
             case 'serial_number':
                 if ($filters->field_filter == Objects::FILTER_SERIAL) {
                     $process = true;
-                    continue;
                 }
                 break;
             case 'dimension':
                 if ($filters->field_filter == Objects::FILTER_DIM) {
                     $process = true;
-                    continue;
                 }
                 break;
             case 'object_id':
                 if ($filters->field_filter === Objects::FILTER_ID) {
                     $process = true;
-                    continue;
                 }
                 break;
         }
@@ -584,7 +588,7 @@ class LendObject
     }
 
     /**
-     * Delete object
+     * Delete object an all his rents
      *
      * @return boolean
      */
@@ -594,7 +598,6 @@ class LendObject
             $delete = $this->zdb->delete(LEND_PREFIX . self::TABLE)
                     ->where(array(self::PK => $this->object_id));
             $this->zdb->execute($delete);
-            return true;
         } catch (\Exception $e) {
             Analog::log(
                 'Something went wrong :\'( | ' . $e->getMessage() . "\n" .
@@ -603,6 +606,7 @@ class LendObject
             );
             throw $e;
         }
+return $this->object_id;
     }
 
     /**
