+++ lib/GaletteObjectsLend/Repository/Preferences.php
@@ -39,7 +39,7 @@
  * @since     Available since 0.7
  */
 
-namespace GaletteObjectsLend;
+namespace GaletteObjectsLend\Repository;
 
 use Analog\Analog;
 use \Zend\Db\Sql\Predicate;
